brain_forensics/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
nltk==3.8.1
textblob==0.17.1
matplotlib==3.7.2
networkx==3.1
seaborn==0.12.2
python-dotenv==1.0.0
pymongo==4.5.0
fpdf2==2.7.4
wordcloud==1.9.2
scikit-learn==1.3.0
plotly==5.15.0
dash==2.13.0
joblib==1.3.1 